Improved AI and ranged units.
AI now is random, with a few bells and whistles. Ranged units are now in
the game (currently the Star Cruiser is ranged for testing). This is a
simplified version of ranged units that only allows units to attack
without taking damage. They still need to be in "melee" range. This is
also towards facilitating buildings. #17 #10